Create parsing
This script uses the BeautifulSoup library to parse a hypothetical blog post. It extracts the title, author, and content of the blog post from a given URL.

Make sure to install the BeautifulSoup library before running this code:

bash

pip install beautifulsoup4

Note: The HTML structure and class names used in this example are hypothetical. You may need to inspect the actual HTML structure of the page you want to parse and adjust the code accordingly.